backend/app/db/schemas.py
from pydantic import BaseModel
import uuid
from datetime import datetime, date

class user(BaseModel):
    username: str
    password: str

class token(BaseModel):
    access_token: str
    token_type: str