pyproject.toml
[project]
name = "t13-fastapi-agents-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "chainlit>=2.8.2",
    "fastapi[standard]>=0.116.2",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.25",
]
